Rename visitor to `AgentFileVisitor`.
